pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proto-whatsapp-saurus"
version = "0.1.0"
description = "WhatsApp AI Agent with LangGraph"
dependencies = [
    "flask>=2.3.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.5.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.8.0",
    "resend>=0.6.0",
    "redis>=5.0.0",
    "rq>=1.15.0",
]

[tool.setuptools]
py-modules = ["main", "models", "services", "ai_agent"]
